Parsing page data
Now parsing the page to receive the important text inside.
Tested on queries of Binyamin Netanyahu. It works quite well and most of
the time the relevant data is returned without too much trash
It's a bit performance heavy since it involves a lot of parsing of the
page and parent elements, so maybe some optimization will have to be
done (or transfer it to different gears?)

How to run: rabbitmq.py -Q google_in -m "Binyamin Netanyahu" (This is
just for debugging, rabbitmq.py is deprecated)